# docs/swagger.yaml
basePath: /api
definitions:
  handler.AdjustBalanceRequest:
    properties:
      delta:
        example: 100
        type: integer
      remark:
        example: 管理员充值
        type: string
    required:
    - delta
    type: object
...
  handler.LoginRequest:
    properties:
      password:
        example: "123456"
        type: string
      username:
        example: admin
        type: string
    required:
    - password
    - username
    type: object
  handler.PageData:
    properties:
      list: {}
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
    type: object
  handler.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  handler.Verify2FARequest:
    properties:
      code:
        example: "123456"
        type: string
      pre_auth_token:
        type: string
      user_id:
        example: 1
        type: integer
    required:
    - code
    - pre_auth_token
    - user_id
    type: object
  handler.serviceAdjustBalanceReq:
    properties:
      delta:
        type: integer
      remark:
        type: string
    required:
    - delta
    type: object
  handler.serviceConsumeReq:
    properties:
      amount:
        type: integer
      key:
        type: string
    required:
    - key
    type: object
host: localhost:8080
info:
  contact: {}
  description: 卡密管理系统 API 文档
  title: CloudKey API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: 用户名密码验证，成功后返回 pre_auth_token 用于后续 TOTP 验证
      parameters:
      - description: 登录参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: '成功: data 含 require_totp/user_id/pre_auth_token'
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 用户名或密码错误
          schema:
            $ref: '#/definitions/handler.Response'
        "403":
          description: 账号已被锁定
          schema:
            $ref: '#/definitions/handler.Response'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 用户名密码登录
      tags:
      - 认证
  /auth/totp/confirm-init:
    post:
      consumes:
      - application/json
      description: 首次登录用户确认 TOTP 绑定后自动登录，返回 JWT
      parameters:
      - description: 验证参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功并返回 JWT
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    role:
                      type: string
                    tenant_id:
                      type: integer
                    token:
                      type: string
                    token_type:
                      type: string
                    username:
                      type: string
                  type: object
              type: object
        "400":
          description: 验证码错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: pre_auth_token 无效
          schema:
            $ref: '#/definitions/handler.Response'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 确认 TOTP 绑定并登录
      tags:
      - 认证
  /auth/totp/setup-init:
    post:
      consumes:
      - application/json
      description: 首次登录用户生成 TOTP 密钥，需先通过密码验证获取 user_id
      parameters:
      - description: 用户ID
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 生成成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    secret:
                      type: string
                    url:
                      type: string
                  type: object
              type: object
        "401":
          description: 用户不存在或 TOTP 已设置
          schema:
            $ref: '#/definitions/handler.Response'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/handler.Response'
      summary: 初始化 TOTP 绑定
      tags:
      - 认证
  /auth/verify-2fa:
    post:
      consumes:
      - application/json
      description: 登录成功后使用 pre_auth_token + TOTP 验证码完成认证，返回 JWT
      parameters:
      - description: 验证参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.Verify2FARequest'
      produces:
      - application/json
      responses:
        "200":
          description: 验证成功，返回 JWT
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    role:
                      type: string
                    tenant_id:
                      type: integer
                    token:
                      type: string
                    token_type:
                      type: string
                    username:
                      type: string
                  type: object
              type: object
        "401":
          description: 验证码错误或 pre_auth_token 无效
          schema:
            $ref: '#/definitions/handler.Response'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/handler.Response'
      summary: TOTP 二次验证
      tags:
      - 认证
  /service/balance-logs:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页流转日志
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询额度流转日志
      tags:
      - 服务账号API
  /service/balance-logs/export:
    get:
      parameters:
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号导出额度流转日志
      tags:
      - 服务账号API
  /service/consume-result:
    get:
      description: 通过 X-Service-Key 认证，根据 request_id 查询消费或调额的结果
      parameters:
      - description: 请求ID
        in: query
        name: request_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 操作结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 缺少 request_id 参数
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 未找到对应的操作记录
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 根据请求ID查询操作结果
      tags:
      - 服务账号API
  /service/keys:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: '状态过滤: active/exhausted/disabled/expired'
        in: query
        name: status
        type: string
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页卡密列表
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密列表
      tags:
      - 服务账号API
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，服务账号创建卡密
      parameters:
      - description: 卡密参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号创建卡密
      tags:
      - 服务账号API
  /service/keys/{id}:
    delete:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号删除卡密
      tags:
      - 服务账号API
    get:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 卡密详情
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密详情
      tags:
      - 服务账号API
    patch:
      consumes:
      - application/json
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号更新卡密
      tags:
      - 服务账号API
  /service/keys/{id}/adjust-balance:
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，增加或减少卡密额度，所有变动记录在流转日志中
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 调整参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.serviceAdjustBalanceReq'
      produces:
      - application/json
      responses:
        "200":
          description: 调整结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号调整卡密额度
      tags:
      - 服务账号API
  /service/keys/{id}/disable:
    patch:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 禁用成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号禁用卡密
      tags:
      - 服务账号API
  /service/keys/{id}/enable:
    patch:
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 启用成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 无效的卡密 ID
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号启用卡密
      tags:
      - 服务账号API
  /service/keys/consume:
    post:
      consumes:
      - application/json
      description: 通过 X-Service-Key 认证，扣减指定卡密的剩余额度
      parameters:
      - description: 扣减参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.serviceConsumeReq'
      produces:
      - application/json
      responses:
        "200":
          description: 扣减结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误或卡密无效
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号扣减卡密额度
      tags:
      - 服务账号API
  /service/keys/export:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号导出卡密（文本格式）
      tags:
      - 服务账号API
  /service/keys/export/json:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据 JSON 数组
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号导出卡密（JSON 格式）
      tags:
      - 服务账号API
  /service/keys/status:
    get:
      description: 通过 X-Service-Key 认证，根据卡密值查询状态
      parameters:
      - description: 卡密值
        in: query
        name: sk
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 卡密状态信息
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 缺少卡密参数
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 服务账号密钥无效
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 卡密不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ServiceKeyAuth: []
      summary: 服务账号查询卡密状态
      tags:
      - 服务账号API
  /super/configs:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 配置列表
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取系统配置
      tags:
      - 系统管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 配置数组
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新系统配置
      tags:
      - 系统管理
  /super/login-logs:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页登录日志
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 登录日志
      tags:
      - 系统管理
  /super/password:
    put:
      consumes:
      - application/json
      parameters:
      - description: 密码参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 修改成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 旧密码错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 修改密码
      tags:
      - 认证
  /super/profile:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 获取当前用户信息
      tags:
      - 认证
  /super/tenants:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 租户列表
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 租户列表
      tags:
      - 系统管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 租户参数
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 创建成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    admin_password:
                      type: string
                    admin_username:
                      type: string
                    tenant:
                      type: object
                  type: object
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 创建租户
      tags:
      - 系统管理
  /super/tenants/{id}:
    get:
      parameters:
      - description: 租户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 租户详情
          schema:
            $ref: '#/definitions/handler.Response'
        "404":
          description: 租户不存在
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 租户详情
      tags:
      - 系统管理
    patch:
      consumes:
      - application/json
      parameters:
      - description: 租户ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新字段
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 更新租户
      tags:
      - 系统管理
  /super/tenants/{id}/reset-password:
    patch:
      parameters:
      - description: 租户ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 重置成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    new_password:
                      type: string
                  type: object
              type: object
        "400":
          description: 无效的ID
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 重置租户管理员密码
      tags:
      - 系统管理
  /super/totp/confirm:
    post:
      consumes:
      - application/json
      description: 已登录用户输入验证码确认 TOTP 绑定
      parameters:
      - description: 验证码
        in: body
        name: body
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 绑定成功
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 验证码错误
          schema:
            $ref: '#/definitions/handler.Response'
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 确认 TOTP 绑定
      tags:
      - 认证
  /super/totp/setup:
    post:
      description: 已登录用户生成新的 TOTP 密钥，用于更换验证器
      produces:
      - application/json
      responses:
        "200":
          description: 生成成功
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  properties:
                    secret:
                      type: string
                    url:
                      type: string
                  type: object
              type: object
        "401":
          description: 未认证
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 生成 TOTP 密钥
      tags:
      - 认证
  /tenant/balance-logs:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 20
        description: 每页数量
        in: query
        name: page_size
        type: integer
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 分页流转日志
          schema:
            allOf:
            - $ref: '#/definitions/handler.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PageData'
              type: object
      security:
      - ApiKeyAuth: []
      summary: 额度流转日志
      tags:
      - 租户-额度流转日志
  /tenant/balance-logs/export:
    get:
      parameters:
      - description: 别名前缀搜索
        in: query
        name: alias
        type: string
      - description: 后缀精准搜索
        in: query
        name: key_suffix
        type: string
      - description: 开始时间
        in: query
        name: start_time
        type: string
      - description: 结束时间
        in: query
        name: end_time
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 导出数据
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 导出额度流转日志
      tags:
      - 租户-额度流转日志
...
  /tenant/keys/{id}/adjust-balance:
    post:
      consumes:
      - application/json
      description: 管理员增加或减少卡密额度，仅支持增量操作，所有变动记录在流转日志中
      parameters:
      - description: 卡密ID
        in: path
        name: id
        required: true
        type: integer
      - description: 调整参数
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handler.AdjustBalanceRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 调整结果
          schema:
            $ref: '#/definitions/handler.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/handler.Response'
      security:
      - ApiKeyAuth: []
      summary: 调整卡密额度
      tags:
      - 租户-卡密管理
...
securityDefinitions:
  ApiKeyAuth:
    description: 'Bearer JWT token, 格式: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
  ServiceKeyAuth:
    description: 服务账号密钥
    in: header
    name: X-Service-Key
    type: apiKey
swagger: "2.0"