# internal/router/router.go
package router

import (
	_ "CloudKey/docs"
	"CloudKey/internal/errcode"
	"CloudKey/internal/handler"
	"CloudKey/internal/middleware"
	"CloudKey/internal/service"
	"CloudKey/internal/web"
	"io"
	"io/fs"
	"net/http"
	"strings"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/redis/go-redis/v9"
	swaggerFiles "github.com/swaggo/files"
	ginSwagger "github.com/swaggo/gin-swagger"
	"gorm.io/gorm"
)

// serveIndexHTML 从嵌入 FS 读取 index.html 并直接写入响应，绕过 http.FileServer。
// http.FileServer 会把 "/index.html" 重定向到 "./"（301），导致循环。
var indexModTime = time.Now()

func serveIndexHTML(c *gin.Context, fsys fs.FS) {
	f, err := fsys.Open("index.html")
	if err != nil {
		c.Status(http.StatusInternalServerError)
		return
	}
	defer f.Close()
	c.Header("Content-Type", "text/html; charset=utf-8")
	c.Status(http.StatusOK)
	http.ServeContent(c.Writer, c.Request, "index.html", indexModTime, f.(io.ReadSeeker))
}

func SetupRouter(
	authHandler *handler.AuthHandler,
	superHandler *handler.SuperHandler,
	tenantKeyHandler *handler.TenantKeyHandler,
	tenantSAHandler *handler.TenantServiceAccountHandler,
	tenantStatsHandler *handler.TenantStatsHandler,
	tenantUsageLogHandler *handler.TenantUsageLogHandler,
	tenantBalanceLogHandler *handler.TenantBalanceLogHandler,
	serviceBalanceLogHandler *handler.ServiceBalanceLogHandler,
	jwtSecret string,
	db *gorm.DB,
	saSvc *service.ServiceAccountService,
	rdb *redis.Client,
	debug bool,
) *gin.Engine {
	r := gin.New()
	r.Use(gin.Recovery())
	r.Use(middleware.RequestID())
	r.Use(middleware.RequestLogger())
	r.Use(middleware.CORSMiddleware())

	// 禁用 Gin 自动尾斜杠重定向，避免根路径 301 循环
	r.RedirectTrailingSlash = false

	// Swagger UI (仅 debug 模式启用)
	if debug {
		r.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
	}

	// 静态文件服务（React SPA）
	distFS, distErr := fs.Sub(web.FS, "dist")
	if distErr != nil {
		panic("failed to open embedded dist: " + distErr.Error())
	}
	fileServer := http.FileServer(http.FS(distFS))

	// 首页：直接返回 index.html，绕过 http.FileServer 的 "/index.html" → "./" 重定向
	r.GET("/", func(c *gin.Context) {
		serveIndexHTML(c, distFS)
	})

	r.NoRoute(func(c *gin.Context) {
		path := c.Request.URL.Path
		if strings.HasPrefix(path, "/api") {
			handler.NotFound(c, errcode.CodeRouteNotFound, errcode.GetMessage(errcode.CodeRouteNotFound))
			return
		}
		// 尝试提供静态资源，不存在则 fallback 到 index.html（SPA 路由）
		trimmed := strings.TrimPrefix(path, "/")
		if f, err := distFS.Open(trimmed); err == nil {
			f.Close()
			fileServer.ServeHTTP(c.Writer, c.Request)
		} else {
			serveIndexHTML(c, distFS)
		}
	})

	api := r.Group("/api")

	// ========== 认证（公开） ==========
	{
		authLimiter := middleware.RateLimitMiddleware(rdb, middleware.DefaultRateLimitWindow, middleware.DefaultRateLimitMaxRequests)
		api.POST("/auth/login", authLimiter, authHandler.Login)
		api.POST("/auth/verify-2fa", authLimiter, authHandler.Verify2FA)
		api.POST("/auth/totp/setup-init", authLimiter, authHandler.SetupTOTPPublic)
		api.POST("/auth/totp/confirm-init", authLimiter, authHandler.ConfirmTOTPPublic)
	}

	// ========== 系统管理员 ==========
	super := api.Group("/super")
	super.Use(middleware.AuthMiddleware(jwtSecret))
	super.Use(middleware.RequireSuperAdmin())
	{
		super.GET("/tenants", superHandler.ListTenants)
		super.POST("/tenants", superHandler.CreateTenant)
		super.GET("/tenants/:id", superHandler.GetTenant)
		super.PATCH("/tenants/:id", superHandler.UpdateTenant)
		super.PATCH("/tenants/:id/reset-password", superHandler.ResetPassword)

		super.GET("/configs", superHandler.GetConfigs)
		super.PUT("/configs", superHandler.UpdateConfigs)

		super.GET("/profile", authHandler.Profile)
		super.PUT("/password", authHandler.ChangePassword)
		super.POST("/totp/setup", authHandler.SetupTOTP)
		super.POST("/totp/confirm", authHandler.ConfirmTOTP)
		super.GET("/login-logs", superHandler.LoginLogs)
	}

	// ========== 租户管理员 ==========
	tenant := api.Group("/tenant")
	tenant.Use(middleware.AuthMiddleware(jwtSecret))
	tenant.Use(middleware.RequireTenantAdmin(db))
	{
		// 个人设置（不拦截 disabled，供前端探测租户状态）
		tenant.GET("/profile", authHandler.Profile)
		tenant.PUT("/password", authHandler.ChangePassword)
		tenant.POST("/totp/setup", authHandler.SetupTOTP)
		tenant.POST("/totp/confirm", authHandler.ConfirmTOTP)
		tenant.GET("/login-logs", tenantUsageLogHandler.LoginLogs)

		// 以下路由受 TenantDisabledGuard 保护（disabled 租户不可访问）
		guarded := tenant.Group("")
		guarded.Use(middleware.TenantDisabledGuard())
		{
			// Key 配置
			guarded.GET("/key-config", tenantKeyHandler.GetKeyConfig)
			guarded.PATCH("/key-config", middleware.TenantBusinessGuard(db), tenantKeyHandler.UpdateKeyConfig)

			// Key 管理（业务操作加 BusinessGuard）
			tenantKeys := guarded.Group("/keys")
			tenantKeys.POST("", middleware.TenantBusinessGuard(db), tenantKeyHandler.CreateKey)
			tenantKeys.GET("", tenantKeyHandler.ListKeys)
			tenantKeys.GET("/status", tenantKeyHandler.Status)
			tenantKeys.GET("/consume-result", tenantKeyHandler.GetConsumeResult)
			tenantKeys.POST("/consume", middleware.TenantBusinessGuard(db), tenantKeyHandler.Consume)
			tenantKeys.GET("/export", tenantKeyHandler.ExportKeys)
			tenantKeys.GET("/export/json", tenantKeyHandler.ExportKeysJSON)
			tenantKeys.GET("/:id", tenantKeyHandler.GetKey)
			tenantKeys.PATCH("/:id", middleware.TenantBusinessGuard(db), tenantKeyHandler.UpdateKey)
			tenantKeys.POST("/:id/adjust-balance", middleware.TenantBusinessGuard(db), tenantKeyHandler.AdjustBalance)
			tenantKeys.PATCH("/:id/disable", middleware.TenantBusinessGuard(db), tenantKeyHandler.DisableKey)
			tenantKeys.PATCH("/:id/enable", middleware.TenantBusinessGuard(db), tenantKeyHandler.EnableKey)
			tenantKeys.DELETE("/:id", middleware.TenantBusinessGuard(db), tenantKeyHandler.DeleteKey)

			// 服务账号（全部加 BusinessGuard）
			tenantSA := guarded.Group("/service-accounts")
			tenantSA.Use(middleware.TenantBusinessGuard(db))
			{
				tenantSA.GET("", tenantSAHandler.ListServiceAccounts)
				tenantSA.POST("", tenantSAHandler.CreateServiceAccount)
				tenantSA.PATCH("/:id/toggle", tenantSAHandler.ToggleServiceAccount)
				tenantSA.DELETE("/:id", tenantSAHandler.DeleteServiceAccount)
			}

			// 统计（不 guard，expired 可查看）
			tenantStats := guarded.Group("/stats")
			{
				tenantStats.GET("/dashboard", tenantStatsHandler.Dashboard)
				tenantStats.GET("/overview", tenantStatsHandler.Overview)
				tenantStats.GET("/trends", tenantStatsHandler.Trends)
				tenantStats.GET("/top-keys", tenantStatsHandler.TopKeys)
				tenantStats.GET("/top-amount", tenantStatsHandler.TopAmount)
				tenantStats.POST("/refresh-top", tenantStatsHandler.RefreshTopStats)
			}

			// 使用日志（不 guard）
			tenantLogs := guarded.Group("/usage-logs")
			{
				tenantLogs.GET("", tenantUsageLogHandler.ListLogs)
				tenantLogs.GET("/export", tenantUsageLogHandler.ExportLogs)
			}

			// 额度流转日志（不 guard）
			tenantBalanceLogs := guarded.Group("/balance-logs")
			{
				tenantBalanceLogs.GET("", tenantBalanceLogHandler.ListLogs)
				tenantBalanceLogs.GET("/export", tenantBalanceLogHandler.ExportLogs)
			}
		}
	}

	// ========== 服务账号 API ==========
	serviceAPI := api.Group("/service")
	serviceAPI.Use(middleware.ServiceAuthMiddleware(saSvc, db))
	{
		serviceAPI.GET("/consume-result", tenantSAHandler.ServiceGetConsumeResult)
	serviceAPI.POST("/keys", middleware.TenantBusinessGuard(db), tenantSAHandler.ServiceCreateKey)
		serviceAPI.GET("/keys", tenantSAHandler.ServiceListKeys)
		serviceAPI.GET("/keys/status", tenantSAHandler.ServiceGetKeyStatus)
		serviceAPI.POST("/keys/consume", tenantSAHandler.ServiceConsumeKey)
		serviceAPI.GET("/keys/export", tenantSAHandler.ServiceExportKeys)
		serviceAPI.GET("/keys/export/json", tenantSAHandler.ServiceExportKeysJSON)
		serviceAPI.GET("/keys/:id", tenantSAHandler.ServiceGetKey)
		serviceAPI.PATCH("/keys/:id", middleware.TenantBusinessGuard(db), tenantSAHandler.ServiceUpdateKey)
		serviceAPI.POST("/keys/:id/adjust-balance", middleware.TenantBusinessGuard(db), tenantSAHandler.ServiceAdjustBalance)
		serviceAPI.PATCH("/keys/:id/disable", middleware.TenantBusinessGuard(db), tenantSAHandler.ServiceDisableKey)
		serviceAPI.PATCH("/keys/:id/enable", middleware.TenantBusinessGuard(db), tenantSAHandler.ServiceEnableKey)
		serviceAPI.DELETE("/keys/:id", middleware.TenantBusinessGuard(db), tenantSAHandler.ServiceDeleteKey)
		serviceAPI.GET("/balance-logs", serviceBalanceLogHandler.ServiceListLogs)
		serviceAPI.GET("/balance-logs/export", serviceBalanceLogHandler.ServiceExportLogs)
	}

	return r
}